Rv and fA range warning tests consolidated
